Updated TR's stat tracker to include long int-1

+++ traffic_router/core/src/main/java/com/comcast/cdn/traffic_control/traffic_router/core/router/StatTracker.java
@@ -332,10 +332,10 @@ public class StatTracker {
 		if(totalHttpCount==0) { return 0; }
 		return totalHttpTime/totalHttpCount;
 	}
-	public int getTotalDsMissCount() {
+	public long getTotalDsMissCount() {
 		return totalDsMissCount;
 	}
-	public void setTotalDsMissCount(final int totalDsMissCount) {
+	public void setTotalDsMissCount(final long totalDsMissCount) {
 		this.totalDsMissCount = totalDsMissCount;
 	}
 
@@ -343,7 +343,7 @@ public class StatTracker {
 	private long totalDnsTime;
 	private long totalHttpCount;
 	private long totalHttpTime;
-	private int totalDsMissCount = 0;
+	private long totalDsMissCount = 0;
 	public Map<String,Long> getUpdateTracker() {
 		return TrafficRouterManager.getTimeTracker();
 	}
